+++ website/common/script/index.js
@@ -137,7 +137,8 @@ api.fns = {
 
 import scoreTask from './ops/scoreTask';
 import sleep from './ops/sleep';
-import allocate from './ops/allocate';
+import allocateNow from './ops/stats/allocateNow';
+import allocate from './ops/stats/allocate';
 import buy from './ops/buy';
 import buyGear from './ops/buyGear';
 import buyHealthPotion from './ops/buyHealthPotion';
@@ -145,7 +146,6 @@ import buyArmoire from './ops/buyArmoire';
 import buyMysterySet from './ops/buyMysterySet';
 import buyQuest from './ops/buyQuest';
 import buySpecialSpell from './ops/buySpecialSpell';
-import allocateNow from './ops/allocateNow';
 import hatch from './ops/hatch';
 import feed from './ops/feed';
 import equip from './ops/equip';
