+++ website/common/script/ops/index.js
@@ -3,7 +3,8 @@ import revive from './revive';
 import reset from './reset';
 import reroll from './reroll';
 import rebirth from './rebirth';
-import allocateNow from './allocateNow';
+import allocate from './stats/allocate';
+import allocateNow from './stats/allocateNow';
 import sortTask from './sortTask';
 import updateTask from './updateTask';
 import deleteTask from './deleteTask';
@@ -35,7 +36,6 @@ import hatch from './hatch';
 import unlock from './unlock';
 import changeClass from './changeClass';
 import disableClasses from './disableClasses';
-import allocate from './allocate';
 import readCard from './readCard';
 import openMysteryItem from './openMysteryItem';
 import scoreTask from './scoreTask';
