+++ website/common/script/ops/stats/allocateNow.js
@@ -1,5 +1,5 @@
 import times from 'lodash/times';
-import autoAllocate from '../fns/autoAllocate';
+import autoAllocate from '../../fns/autoAllocate';
 
 module.exports = function allocateNow (user) {
   times(user.stats.points, () => autoAllocate(user));
